1-kurs/smart_inventory_project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
scikit-learn>=1.4.0
numpy>=1.24.0
pytest==7.4.3
httpx==0.25.2
